+++ pkg/gatewayserver/scheduling/clock.go
@@ -59,9 +59,10 @@ func (c *RolloverClock) SyncTime() (time.Time, bool) {
 // Sync synchronizes the clock with the given concentrator timestamp and the server time.
 func (c *RolloverClock) Sync(timestamp uint32, server time.Time) ConcentratorTime {
 	rollovers := int64(c.absolute/ConcentratorTime(time.Microsecond)) >> 32
-	// If the new timestamp is lower, it is considered a roll over if it advanced more than 25%.
+	// If the new timestamp is lower, it is considered a roll over if it advanced more than 1 minute.
+	// LBS xTime can jump backwards upto 100 ms (https://github.com/TheThingsNetwork/lorawan-stack/issues/2581#issuecomment-672026463).
 	// Otherwise, this is considered an out-of-order arrival.
-	if passed := int64(timestamp) - int64(c.relative); passed < -0x40000000 {
+	if passed := int64(timestamp) - int64(c.relative); passed < -0x3938700 {
 		rollovers++
 	}
 	// If there are full roll over windows between the last and the current server time, take these into account.